argo-platform/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.2,<3.0.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# AI & NLP
langchain>=0.0.350,<0.2.0
langchain-community>=0.0.10,<0.2.0
chromadb>=0.4.18,<0.5.0
sentence-transformers>=2.2.2
openai>=1.6.1
mistralai>=0.0.12
huggingface-hub>=0.19.4
transformers>=4.36.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.4.0

# Streamlit Frontend
streamlit==1.29.0
streamlit-authenticator==0.2.3
plotly==5.17.0
folium==0.15.1
streamlit-folium==0.15.0
geopandas==0.14.1
shapely==2.0.2

# Data Processing
pandas==2.1.4
numpy==1.25.2
xarray==2023.12.0
netcdf4==1.6.5
scipy==1.11.4

# Geospatial
cartopy==0.22.0
rasterio==1.3.9
pyproj==3.6.1

# Caching & Performance
redis==5.0.1
celery==5.3.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1